// model/accountInfo.ts
/**
 * Spendies Backend API
 * This RESTful API is for accessing the back-end functionality of Spendies,  such as registering and verifying accounts, and creating and retrieving wallets, transaction requests, and performing transaction fullfilments.
 *
 * The version of the OpenAPI document: 1.0
 * Contact: connect@spendies.app
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { WalletWithId } from './walletWithId';

export class AccountInfo {
    /**
    * Name of account
    */
    'name': string;
    /**
    * Email address of account
    */
    'email'?: string;
    /**
    * Phone number of account
    */
    'phone'?: string;
    /**
    * Authorisation level of account
    */
    'authLevel': AccountInfo.AuthLevelEnum;
    'wallets'?: Array<WalletWithId>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "email",
            "baseName": "email",
            "type": "string"
        },
        {
            "name": "phone",
            "baseName": "phone",
            "type": "string"
        },
        {
            "name": "authLevel",
            "baseName": "authLevel",
            "type": "AccountInfo.AuthLevelEnum"
        },
        {
            "name": "wallets",
            "baseName": "wallets",
            "type": "Array<WalletWithId>"
        }    ];

    static getAttributeTypeMap() {
        return AccountInfo.attributeTypeMap;
    }
}

export namespace AccountInfo {
    export enum AuthLevelEnum {
        NUMBER_0 = <any> 0,
        NUMBER_1 = <any> 1,
        NUMBER_2 = <any> 2,
        NUMBER_3 = <any> 3
    }
}