// model/errorError.ts
/**
 * Spendies Backend API
 * This RESTful API is for accessing the back-end functionality of Spendies,  such as registering and verifying accounts, and creating and retrieving wallets, payment requests, and performing payment fullfilments.
 *
 * The version of the OpenAPI document: 1.0
 * Contact: connect@spendies.app
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export class ErrorError {
    /**
    * The status code as a number
    */
    'statusCode': number;
    /**
    * The name of the error as a \"bumpy case\", i.e. NotFound or  InternalServerError.
    */
    'name': string;
    /**
    * The traditional error message, which should be kept short and  all single line
    */
    'message'?: string;
    /**
    * The application error code
    */
    'code'?: string;
    /**
    * The application error details
    */
    'details'?: Array<object>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "statusCode",
            "baseName": "statusCode",
            "type": "number"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "message",
            "baseName": "message",
            "type": "string"
        },
        {
            "name": "code",
            "baseName": "code",
            "type": "string"
        },
        {
            "name": "details",
            "baseName": "details",
            "type": "Array<object>"
        }    ];

    static getAttributeTypeMap() {
        return ErrorError.attributeTypeMap;
    }
}